requirements.txt
python-dotenv>=1.0.0
websocket-client>=1.6.0
